[#17] Refactors from PR
